Move InteractiveGraphDescription
