+++ packages/perseus-editor/src/widgets/interactive-graph-editor/components/interactive-graph-description.test.tsx
@@ -4,8 +4,8 @@ import {render, screen} from "@testing-library/react";
 import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
-import {testDependencies} from "../../../../../testing/test-dependencies";
-import InteractiveGraphDescription from "../interactive-graph-description";
+import {testDependencies} from "../../../../../../testing/test-dependencies";
+import InteractiveGraphDescription from "./interactive-graph-description";
 
 import type {UserEvent} from "@testing-library/user-event";
